requirements.txt
peewee==3.15.4
Pillow==9.3.0
pyocr==0.8.3
